+++ server/plugins/server.service.tcpforward/TcpForwardMessenger.cs
@@ -192,7 +192,7 @@ namespace server.service.tcpforward
                 connection.Write(Helper.FalseArray);
                 return ;
             }
-            if (serviceAccessValidator.Validate(connection, EnumServiceAccess.Setting) == false)
+            if (serviceAccessValidator.Validate(connection, (uint)EnumServiceAccess.Setting) == false)
             {
                 connection.Write(Helper.FalseArray);
                 return ;
