+++ server/plugins/server.service.udpforward/UdpForwardMessenger.cs
@@ -162,7 +162,7 @@ namespace server.service.udpforward
                 connection.Write(Helper.FalseArray);
                 return;
             }
-            if (serviceAccessValidator.Validate(connection, EnumServiceAccess.Setting) == false)
+            if (serviceAccessValidator.Validate(connection, (uint)EnumServiceAccess.Setting) == false)
             {
                 connection.Write(Helper.FalseArray);
                 return;
